Add unit tests for bypass marker handling and enhance Sandbox model
- Introduced tests for reading and writing bypass marker files, ensuring correct behavior for various scenarios including missing files, overwriting values, and handling trailing newlines.
- Added functionality to append runtimes while avoiding duplicates and handling blank entries.
- Enhanced the Sandbox struct in the models package with OwnerRef and FleetSuspended fields for better internal bookkeeping and user token scoping.
