+++ pkg/api/server_test.go
@@ -92,7 +92,7 @@ func TestRequireAuthCases(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			server := NewServer(slog.New(slog.NewTextHandler(io.Discard, nil)), nil, nil, config.Config{}, "pat-token")
+			server := NewServer(slog.New(slog.NewTextHandler(io.Discard, nil)), nil, nil, config.Config{}, "pat-token", nil)
 			request := httptest.NewRequest(http.MethodPost, tc.path, strings.NewReader(tc.body))
 			if tc.authorization != "" {
 				request.Header.Set("Authorization", tc.authorization)
@@ -122,7 +122,7 @@ func TestRequireAuthCases(t *testing.T) {
 //  3. GET /v1/metrics is PAT-gated and renders aerolvm_* expvars in
 //     Prometheus text format for production scrapers.
 func TestDashboardEndpoints(t *testing.T) {
-	server := NewServer(slog.New(slog.NewTextHandler(io.Discard, nil)), nil, nil, config.Config{}, "pat-token")
+	server := NewServer(slog.New(slog.NewTextHandler(io.Discard, nil)), nil, nil, config.Config{}, "pat-token", nil)
 
 	t.Run("ui_is_served_unauthenticated_as_html", func(t *testing.T) {
 		request := httptest.NewRequest(http.MethodGet, "/ui", nil)
