+++ pkg/controlplane/controlplane.go
@@ -30,6 +30,38 @@ type Identity struct {
 	OwnerRef   string
 }
 
+// Access is the authenticated caller context attached by the API auth
+// middleware and read by the service layer to enforce owner scoping. Operator
+// is true for the PAT path (operator/admin: bypasses scoping, sees the whole
+// fleet); for a validated user token Operator is false and Identity.OwnerRef is
+// the tenant the request is scoped to.
+//
+// Background loops and internal calls run with no Access in context. The
+// service layer treats "no Access present" as unscoped (fleet-wide) so those
+// paths are never accidentally constrained — only an explicit non-operator
+// Access narrows visibility.
+type Access struct {
+	Identity Identity
+	Operator bool
+}
+
+type accessCtxKey struct{}
+
+// ContextWithAccess returns a child context carrying a. The auth middleware
+// calls this after authenticating; downstream service methods read it back via
+// AccessFromContext.
+func ContextWithAccess(ctx context.Context, a Access) context.Context {
+	return context.WithValue(ctx, accessCtxKey{}, a)
+}
+
+// AccessFromContext returns the Access attached to ctx, if any. ok is false for
+// background loops and internal calls that never went through the auth edge —
+// callers must treat that as unscoped (operator-equivalent) access.
+func AccessFromContext(ctx context.Context) (Access, bool) {
+	a, ok := ctx.Value(accessCtxKey{}).(Access)
+	return a, ok
+}
+
 // Sample is one neutral usage record. It carries units, never money — the
 // managed side is solely responsible for any interpretation of these values.
 type Sample struct {
@@ -73,24 +105,42 @@ type Enforcement interface {
 	Start(ctx context.Context)
 }
 
+// Admitter is the create-admission gate. The daemon consults it before
+// reserving capacity for a new sandbox: a non-nil error refuses the create
+// (e.g. the owner's account is suspended, or fleet standing is unknown beyond
+// the contract grace). The no-op Admitter admits everything, so the
+// open-source build never gates creates.
+type Admitter interface {
+	Admit(ctx context.Context, ownerRef string) error
+}
+
 // Provider bundles the capabilities a build supplies to the daemon. The
 // open-source build uses Noop(); a managed build constructs one backed by the
 // private client. Passed explicitly into the API server and background wiring —
 // there is no global registry, so the dependency stays visible and testable.
+//
+// EnforcementFor is a factory rather than a ready Enforcement because the
+// FleetController it drives is the service layer, which only exists partway
+// through daemon boot. The daemon calls EnforcementFor(svc).Start(ctx) once the
+// service is constructed. A nil factory (or one returning the no-op) means no
+// enforcement loop runs.
 type Provider struct {
-	Validator   Validator
-	Reporter    Reporter
-	Enforcement Enforcement
+	Validator      Validator
+	Reporter       Reporter
+	Admitter       Admitter
+	EnforcementFor func(FleetController) Enforcement
 }
 
 // Noop returns a Provider whose capabilities do nothing: every token is
-// rejected, every report is dropped, and enforcement never runs. This is the
-// open-source default and the safe fallback whenever the feature is disabled.
+// rejected, every report is dropped, every create is admitted, and enforcement
+// never runs. This is the open-source default and the safe fallback whenever
+// the feature is disabled.
 func Noop() Provider {
 	return Provider{
-		Validator:   noopValidator{},
-		Reporter:    noopReporter{},
-		Enforcement: noopEnforcement{},
+		Validator:      noopValidator{},
+		Reporter:       noopReporter{},
+		Admitter:       noopAdmitter{},
+		EnforcementFor: func(FleetController) Enforcement { return noopEnforcement{} },
 	}
 }
 
@@ -104,8 +154,11 @@ func (p Provider) WithDefaults() Provider {
 	if p.Reporter == nil {
 		p.Reporter = noopReporter{}
 	}
-	if p.Enforcement == nil {
-		p.Enforcement = noopEnforcement{}
+	if p.Admitter == nil {
+		p.Admitter = noopAdmitter{}
+	}
+	if p.EnforcementFor == nil {
+		p.EnforcementFor = func(FleetController) Enforcement { return noopEnforcement{} }
 	}
 	return p
 }
@@ -119,6 +172,10 @@ func (noopValidator) Validate(context.Context, string) (Identity, error) {
 type noopReporter struct{}
 
 func (noopReporter) Report(context.Context, []Sample) error { return nil }
+
+type noopAdmitter struct{}
+
+func (noopAdmitter) Admit(context.Context, string) error { return nil }
 
 type noopEnforcement struct{}
